Implemented Alarm Management
Relative (to given time) and absolute (at given time) Alarms can now be set by the ExperimentController via ExperimentAlarmManager. The notification triggers with custom sound. By clicking on it, the user will be directed to the SurveyEntranceActivity. SurveyEntranceActivity is an observable (like every other Activity in this app) and informs the ExperimentController about the event (EVENT_SURVEY_STARTED) and passes it's context. With that the ExperimentController can react to this event by searching the model (currentExperiment) for the next step to follow. Depending on the StepType the ExperimentController fires an explicit intent, fills it's extras with the correct data from the model and starts the corresponding Activity.