InstructionActivity Layout finished
Instructions can now show a picture. If they don't, all the space is used for text. This text scales automatically depending on it's length. The max length of the text is deposited in the Instruction class right above the text attribute. It has to be checked in the Experiment Editor Interface.