InstructionActivity improvement and BreathingExercise implemented
Instructions can now have a duration. That is the case if a instruction should run parallel over the following steps. Steps after this instruction can be modified to wait for it by passing its id (waitForId = instructionToWaitFor.getId()). If the ExperimentController reaches a step that has to wait for this instruction it checks if the timer of this Instruction has already passed. If not the user gets directed to the WaitingRoomActivity. In that the user gets informed, what he has to wait for and that he will be automatically redirected to the next step when the wait time is over.

The BreathingExercise is implemented. It can have to modes, BreathingMode.MOVING_CIRCLE and BreathingMode.STATIC_CIRCLE. It has an instruction screen and a dischargeScreen. Breathing frequency, duration and texts can be set in the experiment editor interface.