TextQuestionActivity implemented / Fixed csv structure issue / optimized ChoiceQuestionActivity
Text Questions are now available.

CsvCreator escapes unwanted characters correctly and writes an "*" at the end of each csv line, because when saving the String csvLine into internal storage all "\n"'s get cut out. "*" acts as a placeholder for "\n" and is replaced with it when the csv file gets loaded in external storage.

ChoiceQuestionActivity shows some hints depending on the visibility of the answers and if multiple answers are allowed.