Fixed Activity Lifecycle
Lifecycle was broken due to the noHistory attributes in the Manifest. Now the user is able to exit the app via the home button and then return after it to the left state.

TODO:
If the user kills the App by swiping it away in the recent apps list the whole Experiment is broken. The only solution to this is implementing a service to get a callback for this situation and then notify the experiment controller to set the next survey alarm.